refactor(test_generation): extract constants, types, and discovery helpers
test_generation.py was 1921 LOC dominated by the AgentTestGenerator class.
Around it sat ~145 LOC of prompt libraries / regex / synthesis-prompt
constants, three small dataclasses, and ~140 LOC of OpenAPI-discovery
free functions — none of which belong wedged between the imports and
the class body or trailing it 1500 lines later.

Extracted into:
- test_generation_constants.py — prompt libraries, refusal patterns,
  tool-keyword lists, regexes, synthesis system prompt (188 LOC)
- test_generation_types.py — ProbeResult, GenerationResult,
  PromptCandidate, _normalize_text_for_comparison (53 LOC)
- test_generation_discovery.py — discover_tool_schemas, OpenAPI
  parsing, _normalize_name, _utc_now_iso, _trace_from_log_entry (158 LOC)

test_generation.py is now 1638 LOC and contains imports +
AgentTestGenerator + run_generation/load_seed_prompts. The class body
itself is the irreducible core; further splitting would require mixin
gymnastics that hurt rather than help readability.

REFUSAL_PATTERNS is re-exported (noqa: F401) — core.model_check_scoring
imports it from evalview.test_generation. Test mock targets like
`evalview.test_generation.AgentTestGenerator._select_synthesis_client`
remain valid.

Zero functional change. 1792 tests pass (same baseline).
